Add Madmate
The Madmate is a member of team Impostors, which has no special abilities.
Functionally, the Madmate is a Crewmate role.
If the Impostors win, the Madmate wins as well.
The Madmate does not know who the Impostors are, and vice versa.
The Madmate cannot fix light because helping the Impostors by
interfering in light out sabotage worens UX.